UGENE-2886
git-svn-id: https://local.ugene.unipro.ru/svn/ugene/trunk@6263 7ca3a678-163c-e011-b84f-001e8c0267ff

+++ src/plugins_3rdparty/variants/src/AssemblySamtoolsMpileup.h
@@ -84,7 +84,7 @@ class CallVariantsTaskSettings{
     int gapSize;
     int window;
     float pvalue1;
-    float pvalue2;
+    double pvalue2;
     float pvalue3;
     float pvalue4;
     float pvalueHwe;
